Initial cleanup of this CoordinateHelper thing, which seems completely overdone for a few lines of utility methods that could live as privates in the Abstract mojo.
